bruin-test/requirements.txt
requests==2.22.0
APScheduler==3.6.0
pyhumps==1.3.1
-f https://s3pypi.mettel-automation.net/igzpackages/index.html
igzpackages==2.0.0
